Use reference counting in RocksDB metadata store (#13309)
### Motivation

RocksDB based implementation of MetadataStore right now can only be opened once, since it requires a single writer and it enforces it through a file lock. 

To make it easier to configure different components which may be creating their own instances of MetadataStore, we should make `RocksDbMetadataStore` to work consistently as the other implementations, by using a static map of instances and reference counting to decide when it can really get closed.